Move assistant to seperate file
